requirements.txt
# Core dependencies
textual>=0.27.0
psutil>=5.9.0
matplotlib>=3.5.0
numpy>=1.22.0
pillow>=9.0.0
pyamdgpu>=0.2.0
pyyaml>=6.0

# Optional dependencies
notify2>=0.3.1    # For system notifications
pytest>=7.0.0     # For running tests
pytest-cov>=3.0.0 # For test coverage
mypy>=0.950       # For type checking
black>=22.3.0     # For code formatting
isort>=5.10.1     # For import sorting
